creating src dir and decomposing
